+++ desktop/panel/database.test.js
@@ -145,165 +145,4 @@ for (const subprocess of RUNNERS) {
       });
     }
   });
-
-  describe('basic cassandra/scylladb tests', () => {
-    test(`runs basic cql query`, async () => {
-      if (process.platform !== 'linux') {
-        return;
-      }
-
-      const connectors = [
-        new DatabaseConnectorInfo({
-          type: 'scylla',
-          database: 'test',
-          username: 'cassandra',
-          password_encrypt: new Encrypt('cassandra'),
-        }),
-      ];
-      const dp = new DatabasePanelInfo();
-      dp.database.connectorId = connectors[0].id;
-      dp.content = 'select broadcast_address from system.local;';
-
-      let finished = false;
-      const panels = [dp];
-      await withSavedPanels(
-        panels,
-        async (project) => {
-          const panelValueBuffer = fs.readFileSync(
-            getProjectResultsFile(project.projectName) + dp.id
-          );
-
-          const v = JSON.parse(panelValueBuffer.toString());
-          expect(v).toStrictEqual([{ broadcast_address: '127.0.0.1' }]);
-
-          finished = true;
-        },
-        { evalPanels: true, connectors, subprocessName: subprocess }
-      );
-
-      if (!finished) {
-        throw new Error('Callback did not finish');
-      }
-    });
-  });
-
-  describe('basic mongodb testdata/documents tests', () => {
-    test('basic test', async () => {
-      if (process.platform !== 'linux') {
-        return;
-      }
-
-      const connectors = [
-        new DatabaseConnectorInfo({
-          type: 'mongo',
-          database: 'test',
-          username: 'test',
-          password_encrypt: new Encrypt('test'),
-          extra: { authenticationDatabase: 'admin' },
-        }),
-      ];
-
-      const dp = new DatabasePanelInfo();
-      dp.database.connectorId = connectors[0].id;
-      dp.content = 'db.test.find({ pageCount: { $gt: 0 } }).toArray()';
-
-      let finished = false;
-      const panels = [dp];
-      await withSavedPanels(
-        panels,
-        async (project) => {
-          const panelValueBuffer = fs.readFileSync(
-            getProjectResultsFile(project.projectName) + dp.id
-          );
-
-          const v = Array.from(JSON.parse(panelValueBuffer.toString())).map(
-            (el) => {
-              delete el._id;
-              return el;
-            }
-          );
-
-          expect(v).toStrictEqual(
-            JSON.parse(
-              fs.readFileSync('testdata/mongo/documents.json').toString()
-            )
-          );
-
-          finished = true;
-        },
-        { evalPanels: true, connectors, subprocessName: subprocess }
-      );
-
-      if (!finished) {
-        throw new Error('Callback did not finish');
-      }
-    }, 15_000);
-
-    test('errors with invalid authenticationDatabase', async () => {
-      if (process.platform !== 'linux') {
-        return;
-      }
-
-      const connectors = [
-        new DatabaseConnectorInfo({
-          type: 'mongo',
-          database: 'test',
-          username: 'test',
-          password_encrypt: new Encrypt('test'),
-          extra: { authenticationDatabase: 'invalid' },
-        }),
-      ];
-
-      const dp = new DatabasePanelInfo();
-      dp.database.connectorId = connectors[0].id;
-      dp.content = 'db.test.find({})';
-
-      const panels = [dp];
-      try {
-        await withSavedPanels(panels, () => {}, {
-          evalPanels: true,
-          connectors,
-          subprocessName: subprocess,
-        });
-      } catch (e) {
-        expect(e.name).toBe('UserError');
-        expect(e.message).toBe('MongoServerError: Authentication failed.\n');
-      }
-    }, 15_000);
-
-    test('errors when not using .toArray() on queries returning multiple objects', async () => {
-      if (process.platform !== 'linux') {
-        return;
-      }
-
-      const connectors = [
-        new DatabaseConnectorInfo({
-          type: 'mongo',
-          database: 'test',
-          username: 'test',
-          password_encrypt: new Encrypt('test'),
-        }),
-      ];
-
-      const dp = new DatabasePanelInfo();
-      dp.database.connectorId = connectors[0].id;
-      dp.content = 'db.test.find({ pageCount: { $gt: 0 } })';
-
-      const panels = [dp];
-      try {
-        await withSavedPanels(panels, () => {}, {
-          evalPanels: true,
-          connectors,
-          subprocessName: subprocess,
-        });
-      } catch (e) {
-        expect(e.name).toBe('UserError');
-        expect(
-          e.message.startsWith(
-            'BSONTypeError: Converting circular structure to EJSON:'
-          )
-        ).toBe(true);
-      }
-    }, 15_000);
-  });
 }
